github: Fix nightly build workflow
twister option '--integration' runs tests for platforms under the
'integration_platforms'. It seems that is the default behavior too.
To run all platforms in 'platform_allow', you need to put '--all'.

Add '--overflow-as-errors' for builds so that any FLASH/RAM overflows
will fail the build.

serial_modem.memfault removed for thingy91 because there is 'FLASH'
overflow by 23656 bytes and the whole thingy91 could be dropped.

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>