app: Buffer URCs during datamode
URCs from modem and other SM threads are buffered during
the data mode. When datamode is exited, all the received
URCs will be received.

Note: Currently this prevents the TCP and UDP clients (and
servers) from receiving data in data mode, as the send is
done from a thread, which is considered URC. There is a
work item to refactor the TCP and UDP clients (and servers),
which will address this.

Signed-off-by: Markus Lassila <markus.lassila@nordicsemi.no>