app: Add #XSENDNTF for when 8192 is set as flags
When 8192 is set for the send operation flags, the
send operation returns immediately with #XSEND marking if
we managed to push the data to the modem.

When modem has pushed the data to network, or (for TCP) the network
has acknowledged the data, #XSENDNTF: <socket>,<status>,<data_sent>
is sent to.

This requires MFW ntn_1.0.0.

Signed-off-by: Markus Lassila <markus.lassila@nordicsemi.no>