app: Remove Zephyr modem driver specific Kconfigs
Zephyr modem driver does not require any special handling so
remove the Kconfigs to support only one operation mode.

When PPP is stopped, it will not restart automatically.
When PPP is stopped, AT channel swiches to DLCI1, if it was somewhere
else.

Recovery of PPP connection should be handled on host side.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>