app: Add CONFIG_SM_AT_ECHO_MAX_LEN to allow truncation of echo
CONFIG_SM_AT_ECHO_MAX_LEN limits the amount of characters that will be
echoed for a singular AT-command when ATE1 is set.

If limit is reached, SM stops echoing the characters, but will echo
the termination characters when those are received.

By default CONFIG_SM_AT_ECHO_MAX_LEN is the size of the
AT-command buffer (4096), so no limitation occurs unless delete or
backspace characters are used while typing. Delete and backspace
are included in the echoed characters but not in the AT-command.

Signed-off-by: Markus Lassila <markus.lassila@nordicsemi.no>