app: Fix a race condition when closing the pipe
Within the UART handler and CMUX there is a possible race-condition
when the pipe is closed.

CMUX module closes the pipe which disables the UART, but immediately
following that the UART is re-opened by calling sm_uart_handler_enable().
This initialization function resets the k_work handlers, which might
have still be in the work queue.

Fix the race condition by:
* Use static initializers for k_work structures.
* Cancel potentially running work handler when closing UART.
* Use k_work_reschedule.. when worker is needed, in case it was running.
* Clear atomic flags before enabling the UART so potentially triggered
  work handler see the up to date values.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>