app: Fix echo with automatic reception and XAPOLL
Automatic reception and XAPOLL do not use the URC-buffer
due to possibly large data amounts that could be received
with automatic reception.

For this reason the worker for processing the events must
only be run when the echo is not in progress, meaning that
either the AT-command has completed or the timer of
CONFIG_SM_URC_DELAY_WITH_INCOMPLETE_ECHO_MS from the last
output has elapsed.

Signed-off-by: Markus Lassila <markus.lassila@nordicsemi.no>