app: Add CMUX modem trace backend
When CMUX trace backend is enabled, Serial Modem will send the
modem traces to the first CMUX channel after the AT-command channel
and the PPP channel. If the channel is not open, traces will be discarded.

Modem trace collection with CMUX backend is enabled with overlay:
overlay-trace-backend-cmux.conf

The UART0 speed should be set to maximum possible value to minimize
the loss of trace data.

Signed-off-by: Markus Lassila <markus.lassila@nordicsemi.no>
Signed-off-by: FT Liang <FT.Liang@garmin.com>