app: Fix race condition to switch CMUX channel for AT
When all UART buffers are empty, and modem pipe buffers are empty,
the "OK" message is immediately send from rsp_send_ok() and if
requeste_at_channel was not set, the AT channel remains at the old
channel, until next URC is send out.

This was causing an URC message to be send to PPP pipe.

Signed-off-by: Seppo Takalo <seppo.takalo@nordicsemi.no>